Revert "Plan 9 from Bell Labs Fourth Edition 2005-12-08"
This reverts commit 8e5b7f3b5f953bdd8dc9e98ff292dbc9e663ff4d.

+++ sys/src/ape/lib/ap/plan9/profile.c
@@ -9,7 +9,7 @@
 extern	long*	_clock;
 extern	long	_callpc(void**);
 extern	long	_savearg(void);
-extern	void*	sbrk(unsigned long);
+extern	void*	sbrk(int);
 
 typedef unsigned long ulong;
